+++ src/main/java/com/nlu/tai_lieu_xanh/application/major/service/AdminMajorService.java
@@ -5,9 +5,9 @@ import com.nlu.tai_lieu_xanh.application.major.dto.request.MajorUpdateRequest;
 import com.nlu.tai_lieu_xanh.application.major.dto.response.MajorResponse;
 
 public interface AdminMajorService {
-  MajorResponse update(Integer MajorId, MajorUpdateRequest request);
+  MajorResponse update(Long MajorId, MajorUpdateRequest request);
 
   MajorResponse save(MajorCreateRequest request);
 
-  void delete(Integer majorId);
+  void delete(Long majorId);
 }
