+++ src/main/java/com/nlu/tai_lieu_xanh/application/user/dto/response/RegisterResponse.java
@@ -3,5 +3,5 @@ package com.nlu.tai_lieu_xanh.application.user.dto.response;
 public record RegisterResponse(
     String message,
     String email,
-    Integer id) {
+    Long id) {
 }
