+++ src/main/java/com/nlu/tai_lieu_xanh/application/user/service/AuthService.java
@@ -21,5 +21,5 @@ public interface AuthService {
 
   void forgotPassword(String email);
 
-  Integer getCurrentUserId();
+  Long getCurrentUserId();
 }
